Database Update. More constants added.
Database.json is no longer tracked! Updates will be done by use of the Database Creator module. I will update the "update database" function in the module and, when updates are given, users will simply run module as is. When completed the update will be done.

This update removes all instances of "Autosubtract" from the database and also removes the need to specify "Subtracted" as a key for each file. New form is:
database[folder][file][data]

Allows easier loops through Database.