Gaussian Fit Guesses and Gaussian Centering Adjusted
Previously, guesses for Gaussian fits were:
convert(np.mean(gaussian))
convert(np.std(gaussian))

This wasn't great because we were actually averaging and finding the SD of the counts, not of the pixels being counted. This has been fixed and has worked on the first test case.

peak_finder method was returning argmax(gaussian) for some reason? I thought this was fixed but maybe it never got pushed? It now returns the mean from the curve_fit so that the Gaussian should be centered better.