find_center,  crop_image, and plot_3d added to Image Class
Minor changes to Image_Processing to accommodate said changes. Plotting.py got its first function plot_3d.

test.py in the main directory has test code validating that the center-finder does in-fact work very consistently. This method is a center-of-mass-style calculation. We'd intended to do this before, but I'm unsure if it was implemented correctly or not. The big test will be the fitting procedure and guessing the SD.

TODO:

Can we bounce b/w pixelspace and distancespace without undoing any spatial shifts that might've happened?

Validate shift_spacial_map method.

Can we plot correctly after the spatial shifts?